Moving TaskInterface and Task to task_scheduler.hpp
This move allows for the removal of the source file, task_scheduler.cpp.

- Updated the documentation in task_scheduler.hpp
- Reorganized FreeRTOS headers in utility/rtos.hpp
