+++ library/L3_Application/L3.mk
@@ -1,2 +1,12 @@
-SOURCES += $(LIBRARY_DIR)/L3_Application/task_scheduler.cpp
+# ==============================================================================
+# Task Scheduler
+# ==============================================================================
+TESTS += $(LIBRARY_DIR)/L3_Application/test/task_scheduler_test.cpp
+TESTS += $(LIBRARY_DIR)/L3_Application/test/periodic_scheduler_test.cpp
+
+# ==============================================================================
+# FILE I/O
+# ==============================================================================
+TESTS += $(LIBRARY_DIR)/L3_Application/file_io/test/fatfs_test.cpp
+
 SOURCES += $(LIBRARY_DIR)/L3_Application/file_io/diskio.cpp
