+++ library/L3_Application/periodic_scheduler.hpp
@@ -4,7 +4,7 @@
 #include "semphr.h"
 #include "timers.h"
 
-#include "L3_Application/task.hpp"
+#include "L3_Application/task_scheduler.hpp"
 #include "utility/enum.hpp"
 #include "utility/log.hpp"
