--- library/L3_Application/test/task_test.cpp
@@ -1,7 +1,0 @@
-#include "L4_Testing/testing_frameworks.hpp"
-#include "L3_Application/task.hpp"
-
-namespace sjsu
-{
-TEST_CASE("Testing Task") {}
-}  // namespace sjsu
